Fully integrated
