+++ src/discord/d.readme.md
@@ -1,0 +1,1 @@
+Welcome to TripSit's Discord bot!
