+++ src/global/utils/combochart.js
@@ -1,0 +1,10 @@
+'use strict';
+
+// const PREFIX = require('path').parse(__filename).name;
+// const logger = require('./logger');
+
+module.exports = {
+  async combochart() {
+    return 'https://i.imgur.com/juzYjDl.png';
+  },
+};
