+++ src/irc/i.readme.md
@@ -1,0 +1,1 @@
+Welcome to TripSit's IRC bot!
