--- src/telegram/commands/index.js
@@ -1,15 +1,0 @@
-'use strict';
-
-/* eslint-disable global-require */
-
-const { Composer } = require('telegraf');
-
-module.exports = Composer.compose([
-  require('./start'),
-  require('./drug'),
-  require('./topic'),
-  require('./combo'),
-  require('./breathe'),
-  require('./combochart'),
-  require('./irc'),
-]);
