+++ src/telegram/commands/t.index.js
@@ -1,0 +1,15 @@
+'use strict';
+
+/* eslint-disable global-require */
+
+const { Composer } = require('telegraf');
+
+module.exports = Composer.compose([
+  require('./t.start'),
+  require('./t.drug'),
+  require('./t.topic'),
+  require('./t.combo'),
+  require('./t.breathe'),
+  require('./t.combochart'),
+  require('./t.irc'),
+]);
