+++ src/telegram/commands/t.topic.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const { Composer } = require('telegraf');
+const { topic } = require('../../global/utils/topic');
+
+module.exports = Composer.command('topic', async ctx => {
+  const data = await topic();
+  ctx.reply(data);
+});
