+++ src/telegram/telegramAPI.js
@@ -2,7 +2,7 @@
 
 const { Telegraf } = require('telegraf');
 const PREFIX = require('path').parse(__filename).name;
-const commands = require('./commands');
+const commands = require('./commands/t.index');
 const logger = require('../global/utils/logger');
 
 const {
@@ -24,7 +24,7 @@ module.exports = {
       'Task failed successfully! 👍', '🤖 TripBot smoked too much pot and fell asleep. Please try again later.', 'Huh, what was that❓ Even my dog can code better! 🐶\nReach out and help us fixing this. :)', "😔 Sorry, i don't know this command.", '🤖 Beep boop beep-- something went wrong.',
     ];
 
-    const m = `${errorMessages.sort(() => 0.5 - Math.random()).slice(0, 30).join(' ')}`;
+    // const m = `${errorMessages.sort(() => 0.5 - Math.random()).slice(0, 30).join(' ')}`;
     bot.on('text', ctx => ctx.reply(errorMessages[Math.floor(Math.random() * errorMessages.length)]));
 
     // Enable graceful stop
