+++ src/webserver/w.readme.md
@@ -1,0 +1,9 @@
+Welcome to TripSit's bot website!
+
+This started because Digital ocean needs a webserver to show that the app is "ready".
+
+So we already have a webserver running, it doesn't hurt to run some static pages
+
+Maybe we can also do some cool stuff while we have the connection to all the bots?
+
+Like maybe the home page can know which channel in each service is least-busy and direct new users there?
